+++ ocky/Observables/MLGame.swift
@@ -63,35 +63,6 @@ class MLGame: NSObject, ObservableObject {
     }
   }
   
-  /* Place in grader/question service */
-  func appendQuestion(question: Question) {
-    self.gameData.history.append(question)
-  }
-  
-  /* Place in grader/question service */
-  func isCorrect(currentQuestion question: Question, usingAnswerChoices choices: [Answer]) -> Bool {
-    let correctAnswers = Set(question.correctAnswers)
-    let userAnswers = Set(choices)
-    let isCorrect = correctAnswers.intersection(userAnswers).count == correctAnswers.count
-    return isCorrect
-  }
-  
-  /* Place in grader/question service */
-  func isCorrect(currentQuestion question: Question, usingAnswerChoices choices: [UUID]) -> Bool {
-    let correctAnswers = Set(question.correctAnswers.map { $0.id })
-    let userAnswers = Set(choices)
-    let isCorrect = correctAnswers.intersection(userAnswers).count == correctAnswers.count
-    return isCorrect
-  }
-  
-  /* Place in grader/question service */
-  func grade(currentQuestion question: Question, usingAnswerChoices choices: [Answer]) {
-    if isCorrect(currentQuestion: question, usingAnswerChoices: choices) {
-      let player = gameData.players.first(where: { $0.displayName == GKLocalPlayer.local.displayName })
-      player?.addCorrectQuestion(id: question.id)
-    }
-  }
-  
   @MainActor
   func loadMatch(matchID: String) async throws {
     do {
@@ -209,11 +180,8 @@ class MLGame: NSObject, ObservableObject {
         self.gameState = state
         self.gameStatePasshtrough.send(state)
       }
-      
     }
-    
   }
-  
 }
 
 extension MLGame {
